requirements.txt
kafka-python==2.0.2
aws-msk-iam-sasl-signer-python==1.0.0
boto3==1.26.125
PyYAML==6.0.1
pytest==8.1.0
python-dotenv==1.0.1